Add files via upload
example of students with courses.
in the program, there are two tables in a database (students and courses). With these tables and the use of hibernate, a new table called enrollment is created. From this created table we make HQL queries to find what is requested.
The courses are randomly selected by the program, for each student, assigning a different number of courses for each student, since that number is determined randomly. finally queries are made and results are printed